backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy==2.0.36
alembic==1.14.1
asyncpg==0.30.0
aiosqlite==0.20.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Validation & Settings
pydantic==2.10.3
pydantic-settings==2.7.1

# HTTP client (for SMS webhook, PLC, etc.)
httpx==0.28.1

# WebSocket
websockets==14.1

# Report generation
reportlab==4.2.5
jinja2==3.1.4

# Date/time
python-dateutil==2.9.0

# Task scheduling (for auto reports)
apscheduler==3.10.4

# Image processing (for OCR, thermal frame)
pillow==11.1.0
numpy==2.2.1

# Logging
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# Dev tools
ruff==0.8.6
black==24.10.0
